requirements.txt
opencv_python==4.4.0.46
Flask==1.1.2
numpy==1.18.5
requests==2.22.0
Pillow==8.0.1
tensorflow==2.3.1
